Linting __version__.
Keeping version in setup.py and __version__ in the base module/package
as hard coded strings. Linting to ensure they remain consistent.

Hardcoding both because setup.py doesn't always have access to read
files in the package, and having the package use pkg_resources to
dynamically get the version is a lot of overhead for __init__.py and
also I once ran across a situation where pip installed a new version of
a package but the contents of the package were from a much older
version. Since package.__version__ just read what pkg_resources said, it
told me it was the latest version when it really wasn't.